+++ source/mouseHandler.py
@@ -9,7 +9,6 @@ import time
 import wx
 import gui
 import tones
-import ctypes
 import winUser
 import queueHandler
 import api
@@ -25,6 +24,7 @@ from math import floor
 from contextlib import contextmanager
 import threading
 from winAPI.winUser.constants import SystemMetrics
+from winBindings import user32
 
 
 WM_MOUSEMOVE = 0x0200
@@ -318,7 +318,7 @@ def getLogicalButtonFlags() -> LogicalButtonFlags:
 	taking into account the Windows user setting
 	for which button (left or right) is primary and which is secondary.
 	"""
-	swappedButtons = ctypes.windll.user32.GetSystemMetrics(SystemMetrics.SWAP_BUTTON)
+	swappedButtons = user32.GetSystemMetrics(SystemMetrics.SWAP_BUTTON)
 	if not swappedButtons:
 		return LogicalButtonFlags(
 			primaryDown=winUser.MOUSEEVENTF_LEFTDOWN,